github-runner-controller-function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-queue
azure-mgmt-containerinstance
azure-identity
azure-containerregistry
azure-keyvault-secrets
azure-storage-blob
azure-mgmt-containerregistry
